refactor folder structure for test files
