+++ src/server/tests/controllers/role.spec.js
@@ -1,7 +1,7 @@
 // import dependencies
 import request from 'supertest'; // import test tool
 import should from 'should'; // import should for accertion
-import server from '../../server/app'; // import the app
+import server from '../../../server/app'; // import the app
 import * as role from '../helpers/roles'; // import role helper
 
 process.env.NODE_ENV = 'test'; // set enviroment to test
